requirements.txt
Django==5.0.2
djangorestframework
drf-yasg
psycopg[binary]>=3.1,<4.0
django-cors-headers
djangorestframework-simplejwt
django-filter
requests
django-oauth-toolkit
django-currentuser
gunicorn
django-otp
celery
redis
django-cron
email-reply-parser
python-dotenv
pillow
channels
channels-redis
django-celery-beat
daphne
django-redis
python-decouple
openpyxl
google-generativeai>=0.7,<1.0
numpy>=2.1,<3.0
pydantic>=2.7.0,<3.0
cryptography>=42.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
msal>=1.29.0
dnspython>=2.6.0
beautifulsoup4